Add files via upload
I implement the UI. I have not yet implemented the feature to store the images in the physical directory and display them on the window, it will be done after connecting the model and UI so that it can convert the video to images run the model, and save to the directory. Delete is working now.  Moving to another folder is able to pull up the windows and let you select the folder, because I cannot save images right now, I am not able to test if this feature working or not